Allow for branch names with slashs in them.
Previously, the assumption was that branch names wouldn't have slashes in
them.  This is incorrect, since branch names in the form:

TA/some-feature

Are quite common.  This fixes that by not splitting on "/" and simply
removing the "refs/head/" prefix from the HEAD file, which then leaves the
full branch name in its entirety.

Signed-off-by: Thomas Adam <thomas.adam22@gmail.com>